+++ packages/core/src/lighting/enums/SSAOQuality.ts
@@ -1,0 +1,11 @@
+/**
+ * Screen Space Ambient Occlusion quality levels.
+ */
+export enum SSAOQuality {
+  /** Low quality - fewer samples, faster performance */
+  Low = 0,
+  /** Medium quality - balanced samples and performance */
+  Medium = 1,
+  /** High quality - more samples, better quality */
+  High = 2
+}
