+++ packages/core/src/lighting/screenSpaceLighting/ScreenSpaceAmbientOcclusion.ts
@@ -2,18 +2,9 @@ import { Shader, ShaderMacro, ShaderPass, ShaderProperty } from "../../shader";
 import blitVs from "../../shaderlib/extra/Blit.vs.glsl";
 import scalableAmbientOcclusionFS from "./shaders/ScalableAmbientOcclusion.glsl";
 import bilateralBlurFS from "./shaders/Blur/BilateralBlur.glsl";
+import { SSAOQuality } from "../enums/SSAOQuality";
 
-/**
- * Screen Space Ambient Occlusion quality levels.
- */
-export enum SSAOQuality {
-  /** Low quality - fewer samples, faster performance */
-  Low = 0,
-  /** Medium quality - balanced samples and performance */
-  Medium = 1,
-  /** High quality - more samples, better quality */
-  High = 2
-}
+
 
 /**
  * Screen Space Ambient Occlusion effect configuration.
